Agent AI Pessoal/gandalf/backend-python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pymongo>=4.8.0
motor>=3.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
gunicorn>=22.0.0
